Pandasified FastKnn and added unit test
readVectors() now calls Utils.read_vectors, which does the same thing the previous Utils call did except with Pandas, and works with its return type; rebuild() is modified to work with Pandas DataFrames, like the one readVectors() now returns; and there's a unit test now. Also, I removed the readMikolov() function and the references to keyTransform, but haven't yet changed it so that it doesn't even give you the option to enter the file format that would call readMikolov().