Switch to ParentService-based URL System
URLs for map resources are now added using a class called ParentService,
whose __getattr__() method has been modified to allow it to dynamically
provide the service for the map in question. Well, actually it provides
a function for representing a particular method of a particular service
of every map, and it takes a kwarg called "map_name" to determine which
one to call. It's complicated.