Add Draft Admin View, Move Templates to New Folder
New admin view deals with the active metric on a map. It will probably
be renamed from 'map admin' to 'metric admin', and the user will be
allowed to add and maintain (and delete) multiple metrics for one map.

Also the naming convention for HTML templates has been standardized, and
they have been moved from the static 'web/' directory to the
'templates/' directory, since these templates should never be served as
static files.

gen_config() no longer adds an option for a metric to "Metrics"; that
work is being delegated to this new 'map admin' service.