Can Now Add Metric Live
There's still a whole bunch of breadcrumb-style print statements, and it
has not been tested on parallel servers yet, but this took way longer
than expected and I'm sleepy.

This works by adding a new flag in MapService (._send_update), which
ParentService checks for. If ParentService notices that this flag has
been raised, it updates the modification time of the meta-config. Since
ParentService also monitors the meta-config's mod time, this should
ParentService on every server instance to call
ParentService.update_maps(), which now has logic that will check the mod
time of each individual *map* config (bear with me), and start up a new
MapService if that mod time is different. Clear as mud?

More work still needs to be done in terms of refactoring and documenting
the incredibly unintuitive system of 'touching' the meta-config file to
signal that one or more of the maps needs to be updated from the file.

This also fixes a bug where build logs and errors were sent to a
directory that was not created in time, which caused an error.