Fix Bugs and Add Documentation for DivergingMetric
There were a few bugs preventing DivergingMetric from being used with
the rest of Cartograph. First, it expected colorCode to be an object
with a .colors attribute; there's no way to pass this directly through
the current JSON-based config file scheme. Now–like
QualitativeMetric–DivergingMetric takes a string and looks for a
ColorBrewer palette of the same name.

There was also an issue where a float was being (attempted to) used as
a list index (to get the appropriate color from the palette).

On this issue, there's still a bug (see added FIXME) where the math is
messed up and the color-finder can't handle an actual value equal to
maxVal.

This commit also adds some documentation. It's pretty repetitive from
QualitativeMetric; some editing and consolidation will be necessary.