Fix SequentialMetric
This commit primarily fixes the two problems that were also present in
DivergingMetric and QualitativeMetric. Namely, the constructor expected
to be passed a palettable colorbrewer palette, which is not possible
under the current config file scheme; also, the colors that it returned
(via the .getColor() method) had components on a scale of 0-255 instead
of on a scale of 0.0-1.0.

The constructor now takes the name of a palettable.colorbrewer palette
as a string, and SequentialMetric.getColor() now returns color tuples of
the correct magnitude.